Move app initialization to run_server function
* Move views to separate module
