+++ openprocurement/auction/gong/includeme.py
@@ -1,9 +1,12 @@
+# -*- coding: utf-8 -*-
 from openprocurement.auction.includeme import _register
 from openprocurement.auction.interfaces import IAuctionsServer
-from openprocurement.auction.gong.views import includeme
 
 
 def kadastralProcedure(components):
     _register(components, 'kadastralProcedure')
     server = components.queryUtility(IAuctionsServer)
-    includeme(server)
+    server.add_url_rule(
+        '/auctions/<auction_doc_id>/<path:path>', 'auctions',
+        server.view_functions['auctions_proxy'], methods=['GET', 'POST']
+    )
